==> code/1_data_collection.ipynb <==
**Imports**

import pandas as pd
pd.set_option('display.max_columns', None)
import requests
from bs4 import BeautifulSoup

#setting up url and request:

#all Phish shows
url_all_shows = 'https://api.phish.net/v5/shows/artist/phish.json?order_by=showdate&apikey=DELETED'
url_set_lists = 'https://api.phish.net/v5/setlists.json?order_by=showdate&apikey=DELETED'

res_all_shows = requests.get(url_all_shows)
res_set_lists = requests.get(url_set_lists)

#confirming appropriate status code (not 404, 502, etc.):
print(res_all_shows.status_code)
print(res_set_lists.status_code)

**Creating DataFrame for Shows**

#creating the soup object for shows:
soup_all_shows = BeautifulSoup(res_all_shows.content, 'lxml')

data_all_shows = res_all_shows.json()

data_all_shows['data'][0]

#building 'shows' data dictionary
shows = []

for show in data_all_shows['data']:
    show_dict = {
        'showid': show['showid'],
        'showyear': show['showyear'],
        'showmonth': show['showmonth'],
        'showdate': show['showdate'],
        'permalink': show['permalink'],
        'setlist_notes': show['setlist_notes'],
        'venue': show['venue'],
        'city': show['city'],
        'state': show['state'],
        'country': show['country'],
        'city': show['city'],
        'artist_name': show['artist_name'],
        'tourid': show['tourid'],
        'tour_name': show['tour_name'],
                }
    shows.append(show_dict)

shows_df = pd.DataFrame(shows)
shows_df.head(3)

shows_df.shape

shows_df.info()

shows_df.shape

shows_df.dropna(inplace = True)

shows_df.to_csv('../data/just_shows.csv')

**Creating DataFrame for Setlists**

#creating the soup object for setlists:
soup_set_lists = BeautifulSoup(res_set_lists.content, 'lxml')

data_set_lists = res_set_lists.json()

data_set_lists['data'][0]

#building 'setlists' data dictionary
setlists = []

for song in data_set_lists['data']:
    song_dict = {
        'showid': song['showid'],
        'showdate': song['showdate'],
        'permalink': song['permalink'],
        'showyear': song['showyear'],
        'uniqueid': song['uniqueid'],
        'meta': song['meta'],
        'setlistnotes': song['setlistnotes'],
        'songid': song['songid'],
        'position': song['position'],
        'transition': song['transition'],
        'set': song['set'],
        'isjam': song['isjam'],
        'isreprise': song['isreprise'],
        'tracktime': song['tracktime'],
        'gap': song['gap'],
        'tourid': song['tourid'],
        'tourname': song['tourname'],
        'song': song['song'],
        'is_original': song['is_original'],
        'venueid': song['venueid'],
        'venue': song['venue'],
        'city': song['city'],
        'state': song['state'],
        'country': song['country'],
        'artist_name': song['artist_name'],
                }
    setlists.append(song_dict)

setlists_df = pd.DataFrame(setlists)
setlists_df.head(3)

setlists_df.shape

setlists_df.info()

setlists_df.to_csv('../data/just_setlists.csv')
