==> code/2_data_cleaning.ipynb <==
## Data Cleaning ##

**Imports**

import pandas as pd
pd.set_option('display.max_columns', None)

import numpy as np
np.random.seed(42)

import matplotlib.pyplot as plt
import seaborn as sns
sns.set(style = "darkgrid")

import datetime as dt

**Reading in Shows' Data**

shows_df = pd.read_csv('../data/just_shows.csv').drop(columns = 'Unnamed: 0')

shows_df.head(3)

shows_df.tail(3)

shows_df.shape

shows_df.columns

**Reading in Setlists' Data**

setlists_df = pd.read_csv('../data/just_setlists.csv').drop(columns = 'Unnamed: 0')

setlists_df.head(3)

setlists_df.tail(3)

setlists_df.shape

setlists_df.columns

**Extracing All setlists from shows_df**

phish_shows = shows_df['showid'].tolist()
len(phish_shows)

phish_shows1 = setlists_df['showid'].unique().tolist()
len(phish_shows1)

#making sure all the shows in setlists_df are in shows_df
shows_df = shows_df[shows_df['showid'].isin(phish_shows1)]

shows_df.shape

#making sure all the shows in shows_df are in setlists_df
setlists_df = setlists_df[setlists_df['showid'].isin(phish_shows)]

setlists_df['showid'].nunique()

setlists_df.shape

**Exploring the Data**

setlists_df.info()
#note, showdate needs to be converted to datetime dtype

setlists_df.isna().sum().sort_values(ascending = False)

**Counting the Number of Shows and Songs**

#the number of unique shows since the beginning:
setlists_df['showid'].nunique()

#the number of unique songs played from setlists_df:
setlists_df['songid'].nunique()

setlists_df.loc[setlists_df['is_original'] == 1, 'songid'].nunique()

**Creating Function to Build column 'setlist'**

show_1251253100 = setlists_df.loc[setlists_df['showid'] == 1251253100]['song'].tolist()
show_1251253100

#https://stackoverflow.com/questions/12293208/how-to-create-a-list-of-lists
def make_setlist_col(df):
    shows = []
    for show in df['showid'].unique():
        show = df.loc[df['showid'] == show]['song'].tolist()
        shows.append(show)
        
    shows_df['setlists'] = shows
    
    return shows_df

make_setlist_col(setlists_df)

shows_df.info()

shows_df['setlists'][1]

**Examining Null Values**

shows_df.isna().sum().sort_values(ascending = False)

shows_df[shows_df['state'].isna()]['country'].value_counts()

- The entries where 'state' is null represents enough of my data that I'm not going to drop the entries. Although I can't easily rename the null 'state' to the appropriate ones, given the number of countries across which they are spread, I am going to rename them to 'Not Available'.

shows_df['state'].fillna('Not Available', inplace = True)

shows_df[shows_df['setlist_notes'].isna()][0::3]

- 'setlist_notes' is null across the date range. As with 'state', I am going to rename the nulls to 'Not Available'

shows_df['setlist_notes'].fillna('Not Available', inplace = True)

**Examining dtypes**

shows_df.dtypes

shows_df['showdate'] = pd.to_datetime(shows_df['showdate'])

shows_df.info()

**Saving to .csv**

shows_df.to_csv('../data/showsandsets.csv', index = False)
