==> code/4_prep_for_modeling.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)
import numpy as np
np.random.seed(42)

import matplotlib.pyplot as plt
import seaborn as sns
sns.set(style = "darkgrid")

from sklearn.preprocessing import OneHotEncoder

df = pd.read_csv('../data/showsandsets.csv')

**Building the List of Songs to Model on:**

#the number of unique songs played:
#setlists_df['songid'].nunique()

#setlists_df.loc[setlists_df['is_original'] == 1, 'songid'].nunique()

setlists = df['setlists'].tolist()

setlists_joined = "".join(setlists)
print(setlists_joined[0:982])

setlists_delim = setlists_joined.replace('][', ', ')
# setlists_delim = setlists_joined.replace('[', ', ')
# setlists_delim = setlists_joined.replace(']', ', ')

#I wish I could do 'jams' or 'intros/outro' to designate a transition

#REVISIT REVISIT REVISIT
#REVISIT REVISIT REVISIT
#REVISIT REVISIT REVISIT

print(setlists_delim[0:982])

setlists_string = setlists_delim.split(', ')
setlists_string[0:5]

#number of all songs played:
len(setlists_string)

#number of all unique songs played:
len(set(setlists_string))

#all unique songs played:
unique_songs = sorted(set(setlists_string))
unique_songs[0::12]

- I want to be able to model on a list of the songs played in order to predict the next song (song N + 1 based on the previous N songs played) so I need to map the song titles to integers.

#https://stackoverflow.com/questions/53801614/map-elements-of-a-list-to-their-index-in-another-list
#from itertools import islice
from itertools import islice

idx_list = {song: index for index, song in enumerate(unique_songs)}
#idx_list #commented out for length

- Applying the index_list mapping to the entire (running) setlist:

encoded_setlists_list = [idx_list[song] for song in setlists_string]
len(encoded_setlists_list)

#first twelve songs played:
encoded_setlists_list[0:12]

#Phish's first show:
df['setlists'][0]

#Phish's first show as an encoded list:
encoded_setlists_list[0:8]

#as calculated in '3_EDA' notebook, the avg. length of a Phish setlist is 19.96:
length = 20
model_sets = []

for song in range(length, len(encoded_setlists_list)):
    
    #selecting a range of 20 songs
    songs = encoded_setlists_list[song - length: song + 1]
    
    #building the setlist from songs
    model_sets.append(songs)

model_sets[0:3]

np.array(model_sets)

len(np.array(model_sets)[1])

#https://numpy.org/doc/stable/reference/generated/numpy.save.html
#https://numpy.org/doc/stable/reference/generated/numpy.load.html#numpy.load

np.save('../data/model_sets.npy', np.array(model_sets), allow_pickle = True)
