==> code/5a_modeling.ipynb <==
## Modeling

**Imports**

import pandas as pd
pd.set_option('display.max_columns', None)
import numpy as np
np.random.seed(42)

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.metrics import accuracy_score

from tensorflow import keras
from tensorflow.keras import layers
from tensorflow.keras.callbacks import EarlyStopping
from tensorflow.keras.layers import Dense, Dropout, Embedding, LSTM
from tensorflow.keras.models import Sequential
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.utils import to_categorical
from tensorflow.keras.wrappers.scikit_learn import KerasRegressor

import matplotlib.pyplot as plt
import seaborn as sns

**Reading in Data**

model_sets = np.load('../data/model_sets.npy')
model_sets.shape

#remember, I am modeling for song n + 1 based on songs n previously played:
X_rr, y_rr = model_sets[:, :-1], model_sets[:, -1] 

X_rr.shape, y_rr.shape

**Train-Test Split**

X_train, X_test, y_train, y_test = train_test_split(X_rr, y_rr, 
                                                    test_size = 0.2, 
                                                    random_state = 42)

print(f'X_train and y_train: {X_train.shape}, {y_train.shape}')
print(f'X_test and y_test: {X_test.shape}, {y_test.shape}')

**Modeling - Second Attempt, with GridSearchCV**

#https://www.tensorflow.org/text/tutorials/text_generation
songs = np.array(model_sets).tolist()
songs = [item for sublist in songs for item in sublist]

vocab = set(songs)
print(f'{len(vocab)} unique songs!')

# Length of the vocabulary in chars
vocab_size = len(vocab)

# The embedding dimension
embedding_dim = 970

# Number of RNN units
rnn_units = 516

#https://www.geeksforgeeks.org/python-keras-keras-utils-to_categorical/
#train labels:
X_train_ohe = np.array([to_categorical(x, num_classes = vocab_size) for x in X_train])
y_train_ohe = to_categorical(y_train, num_classes = vocab_size)

#test labels:
X_test_ohe = np.array([to_categorical(x, num_classes = vocab_size) for x in X_test])
y_test_ohe = to_categorical(y_test, num_classes = vocab_size)

print(f'Train shape: {X_train_ohe.shape}, Test shape: {X_test_ohe.shape}')
print(f'Train shape: {y_train_ohe.shape}, Test shape: {y_test_ohe.shape}')

#https://medium.com/analytics-vidhya/understanding-embedding-layer-in-keras-bbe3ff1327ce
#https://www.tensorflow.org/text/tutorials/text_generation
#https://www.analyticsvidhya.com/blog/2021/08/predict-the-next-word-of-your-text-using-long-short-term-memory-lstm/
#https://stackoverflow.com/questions/55774632/gridsearchcv-randomizedsearchcv-with-lstm

def model_func(vocab_size = len(vocab),
               embedding_dim = 970,
               rnn_units = 516,
               #dropout = 0.10,
               learningrate = 0.01):

    model = Sequential()

    #first layer, embedding:
    model.add(Embedding(vocab_size, embedding_dim))

    #second layer, LSTM:
    model.add(LSTM(rnn_units))

    #final dense layer with softmax activation:
    model.add(Dense(embedding_dim, activation ='softmax'))
    
    #setting optimizer:
    opt = Adam(learningrate)
    
    #compiling:
    model.compile(optimizer = opt, loss = 'categorical_crossentropy', metrics = ['accuracy'])

    return model

rnn = KerasRegressor(build_fn = model_func, batch_size = 128, verbose = 2)

param_grid = {
    'epochs' : [20, 30],
    'rnn_units' : [256, 516],
    #'dropout' : [0.05, 0.10],
    'learningrate': [0.01, 0.05]}

#gridsearch:
gs = GridSearchCV(rnn, param_grid = param_grid, cv = 3)

gs.fit(X_train, y_train_ohe,
       validation_data = (X_test, y_test_ohe),
       verbose = 2)

print(gs.best_score_)
gs.best_params_
